pyproject.toml
[project]
name = "genai-sop-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.124.4",
    "ipywidgets>=8.1.8",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "notebook>=7.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "transformers>=4.57.3",
    "trustcall>=0.0.39",
]
